requirements.txt
aiofiles==0.3.1
httptools==0.0.9
sanic==0.5.4
ujson==1.35
uvloop==0.8.0
websockets==3.3
python-dotenv==0.6.4
click==6.7
asyncpg-0.11.0
